Starting for Sprint 2:
We added a sensor based upon the airquality sensor at Breezometer AQI website.
I updated all the compare functions so the user can specific why compare operator should be used in the compare function (<, >, <=, >=, =).   This can be specified in the configuration file
I cleaned up the documentation and added a Getting Started Guide